# api/openapi.yaml
openapi: 3.1.0
info:
  title: GrantSupport Engine API
  version: 0.1.0-beta.3
  description: >
    GrantSupport is a delegated support-access authentication and cryptographic audit engine.
    It enables multi-tenant SaaS applications to issue cryptographically signed, time-bounded,
    single-use support access tokens to vendor support agents with tamper-evident audit logging.
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://api.grantsupport.local/v1
    description: Production HTTPS Reverse Proxy
  - url: http://127.0.0.1:8080
    description: Local Standalone Development Server
paths:
...
  /.well-known/jwks.json:
    get:
      operationId: getJWKS
      summary: JSON Web Key Set (JWKS)
      description: >
        Returns the public RSA keys (RFC 7517) for downstream API gateways and services to cryptographically verify RS256 JWT tokens issued by GrantSupport.
      security: []
      responses:
        '200':
          description: Public JSON Web Key Set
          headers:
            Cache-Control:
              schema:
                type: string
                example: public, max-age=3600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWKSResponse'
        '503':
          $ref: '#/components/responses/RFC7807InternalError'
...
  /api/v1/auth/support/login:
    post:
      operationId: supportLogin
      summary: Claim Support Grant & Login
      description: >
        Called by a vendor support agent to consume a single-use support grant token.
        Atomically invalidates the token, emits an append-only audit event, and issues an RS256 JWT access token whose expiry matches the duration originally requested when the grant was created (durationMinutes, 1–1440 minutes).
        Rate limited to 10 attempts per minute per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportLoginRequest'
      responses:
        '200':
          description: Support grant claimed and session JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportLoginResponse'
        '400':
          $ref: '#/components/responses/RFC7807BadRequest'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '429':
          $ref: '#/components/responses/RFC7807TooManyRequests'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
...
  /api/v1/auth/support/logout:
    post:
      operationId: supportLogout
      summary: Support Agent Logout
      description: >
        Called by an authenticated support agent to voluntarily terminate their active session by invalidating their token version in the revocation store.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Support agent session logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
  /api/v1/auth/support/sessions:
    get:
      operationId: getActiveSessions
      summary: List Active Support Sessions
      description: >
        Called by a customer administrator to retrieve all currently active, redeemed support sessions for their tenant.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of active support sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveSessionsResponse'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
  /api/v1/auth/support/sessions/{grantId}:
    delete:
      operationId: terminateSession
      summary: Terminate Support Session
      description: >
        Called by a customer administrator to immediately expire a specific active support grant and terminate the agent's active session.
      security:
        - BearerAuth: []
      parameters:
        - name: grantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Unique identifier of the support grant session to terminate.
      responses:
        '200':
          description: Support session terminated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        '400':
          $ref: '#/components/responses/RFC7807BadRequest'
        '401':
          $ref: '#/components/responses/RFC7807Unauthorized'
        '500':
          $ref: '#/components/responses/RFC7807InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 signed JSON Web Token carrying tenant and role claims.
  schemas:
...
    JWK:
      type: object
      required:
        - kty
        - use
        - alg
        - n
        - e
      properties:
        kid:
          type: string
          description: Unique Key ID identifying the asymmetric signing key
          example: grantsupport-primary
        kty:
          type: string
          example: RSA
        use:
          type: string
          example: sig
        alg:
          type: string
          example: RS256
        n:
          type: string
          description: Base64URL-encoded RSA modulus parameter
        e:
          type: string
          description: Base64URL-encoded RSA exponent parameter
    JWKSResponse:
      type: object
      required:
        - keys
      properties:
        keys:
          type: array
          items:
            $ref: '#/components/schemas/JWK'
...
    SupportLoginRequest:
      type: object
      required:
        - token
        - agentId
      properties:
        token:
          type: string
          description: The high-entropy single-use support token provided by the customer administrator.
          example: 550e8400-e29b-41d4-a716-446655440000_9f83a8b9487c6e12e2057639f28d8442c8d2345e89a1b2c3d4e5f6a7b8c9d0e1
        agentId:
          type: string
          format: uuid
          description: Unique identifier of the authenticated vendor support engineer claiming the session.
          example: 7f4c935b-16d7-4f9e-a8f2-39c4a852b719
    SupportLoginResponse:
      type: object
      required:
        - success
        - message
        - institution_id
        - access_token
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Support agent authenticated successfully.
        institution_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        access_token:
          type: string
          description: RS256 signed JWT for the SUPPORT_AGENT role with exp dynamically set to match the original grant duration.
          example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
    ActiveSession:
      type: object
      required:
        - grant_id
        - institution_id
        - granted_by_id
        - used_by_id
        - scope
        - used_at
        - expires_at
        - remaining_seconds
      properties:
        grant_id:
          type: string
          format: uuid
          example: 01a003f4-5678-7123-9abc-def012345678
        institution_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        granted_by_id:
          type: string
          format: uuid
          example: 330e8400-e29b-41d4-a716-446655440000
        used_by_id:
          type: string
          format: uuid
          example: 7f4c935b-16d7-4f9e-a8f2-39c4a852b719
        scope:
          type: string
          example: BILLING_READ_ONLY
        whitelisted_ips:
          type: array
          items:
            type: string
          example: ["198.51.100.4"]
        used_at:
          type: string
          format: date-time
          example: 2026-08-15T12:00:00Z
        expires_at:
          type: string
          format: date-time
          example: 2026-08-15T13:00:00Z
        remaining_seconds:
          type: integer
          example: 3540
    ActiveSessionsResponse:
      type: object
      required:
        - success
        - sessions
      properties:
        success:
          type: boolean
          example: true
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/ActiveSession'
...
    StandardSuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully.
...
    RFC7807ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          example: https://grantsupport.io/errors/validation-failed
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: VALIDATION_FAILED
        status:
          type: integer
          description: The HTTP status code generated by the origin server.
          example: 400
        detail:
          type: string
          description: A sanitized human-readable explanation specific to this occurrence of the problem.
          example: The request payload failed schema validation constraints.
        instance:
          type: string
          description: A URI reference identifying the specific occurrence of the problem.
          example: /api/v1/auth/support/grant
        code:
          type: string
          description: Machine-readable application error code.
          example: ERR_INVALID_INPUT
        correlation_id:
          type: string
          description: Unique correlation identifier for tracing.
          example: 01a003f4-5678-7123-9abc-def012345678
        validation_errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: durationMinutes
              message:
                type: string
                example: durationMinutes must be at least 1
  responses:
    RFC7807BadRequest:
      description: Bad Request (Invalid input or constraint violation)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/RFC7807ProblemDetails'
    RFC7807Unauthorized:
      description: Unauthorized (Invalid, expired, or revoked token)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/RFC7807ProblemDetails'
    RFC7807TooManyRequests:
      description: Too Many Requests (Rate limit exceeded)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/RFC7807ProblemDetails'
    RFC7807InternalError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/RFC7807ProblemDetails'